Refactor
Add hook to fill in part of 'struct client' as root in the main server
process before passing it on to the unprivileged helper process which
fills in the rest.

On Darwin we fill in the entire struct in the unprivileged helper
process. On Linux we will also do that. But on other BSDs we will fill
in the UID from a sysctl call that needs to be done as root.