Give clear error message if not started as root
The server is designed to be started as root and to drop privileges
later. We can't function at all if we're started as a less privileged
user, and the user will soon see random error messages about failed
syscalls. Better to check for root privileges right off the bat and
state our requirements clearly.