Use purposed int types instead of uint64_t in IPC
Previously I tried to be conservative and use uint64_t to ship
everything around. This would be a fine idea if the main and child
processes came from separate executables that could be compiled
separately. But since they are same executable, we can trust that the
purpose-built int types like uid_t are the same size in all those
processes. So it's simpler to use the target type right off the bat.
This ought to solve one Coverity defect too.