Remove redundant client uid/gid variables
Another place where I was being overly cautious. In the main process,
we can keep using the struct we read from the helper process even
after we have read the struct. We don't need to "evacuate" data from
the struct into separate variables for fear of the struct being
overwritten. Was this doubt lingering from an experiment where the IPC
was based on shared memory?