Distinguish between root and "user not found"
Use -1 to indicate a missing user (and group) in the IPC
struct. Previously we used zero. Since zero is also the uid of root,
this resulted in the confusing error message "user not found" when
root tried to connect to us. The behavior was still correct, since
root is not allowed to connect anyway, but the message was weird.