Always run ~/.usertcp instead of command given by admin
I thought this through by trying to type out a few scenarios where a
sysadmin would instruct a user on how to set up a private website for
themselves. I also thought about cases where the user's home directory
is on a NFS share so it's accessed from different machines. And
logging. And a couple other things. I came to the conclusion that it's
simpler and better in every way to have the command be always
constant.