Implement listening on multiple ports at once
This is a big change. As many ports as you like can now be given on on
the command line. The server uses poll() to accept clients on all of
them.

poll() is now in Posix so there's no need to bother with the old
select() API and its portability problems. We also don't need the
speed of libev and the like. It's nice when we can keep things simple.